Add handling of multiple schema files
Add read-schemas, to read all .yaml files in a directory as schemas.

Update schema handling in hunchentoot startup by replacing the :schema
parameter with :schemapath, and apply inject-schema to its output.
This replaces the previous behaviour of expecting to be given a single,
already-parsed, schema.

Add generic schema files, both to demonstrate this behaviour, and to
make this thing more useful by default.