Refactor get-relationship-attrs, delete-relationship-by-path
Define a new struct `relationship-attrs`, to be returned by
get-relationship-attrs and thus remove confusion about what information
is to be found where in its return value.
Refactor get-relationship-attrs to return same, and update all
calls to it to properly handle the new return values.

Refactor and bugfix delete-relationship-by-path while we're at it.
Nominally, this is a separate thing but in practice it's not feasible to
separate the changes.