Start baking in a core schema
Refactor the schema-files into the core schema that parts of the
application assume to be there, and an optional set that's useful as a
demo and some inspiration.

Then add the code to pre-compile the core schema, to make sure it's
present in the saved image, and inject it along with any user-defined
schemas that are provided.