Update add-rel-to-schema-rtype for new struct types
Split the existing method into two implementations, according to the
`relationship` parameter.
Specialise one implementation on `incoming-rels`, and the other on
`schema-rels`.

Do this because this method will be applied when creating new
resourcetypes, and will then be applied again if the new resourcetype
turns out to be a duplicate. In the first case, the relationships are
still unfiltered (and thus untrusted) input, usually from a YAML file.
In the second case, it's come through that process already, so the
relationships have been turned into `schema-rels` structs, and now most
of the sanity-checking can be assumed to have been done.