Finally fix the _ vs -- encoding/decoding issue
Control the Lisp<-->JSON encoding/decoding to reliably get the
round-trip result that's most useful in this codebase: downcase all the symbols
on the way out, and upcase them on the way back in.

This ensures that Neo4j receives keywords like 'statements' in
lower-case, as required, and that we robustly ensure that all Neo4j
labels are converted to :UPPERCASE format as expected in traditional CL
style.

This should permanently end a long-running series of WTFs that keep
breaking the test suite, as well as ensure we don't get other unexpected
behaviour of this kind.