Define *hunchentoot-acceptor* in config.lisp
Move it out of `hunchentoot.lisp` to stop the compiler complaining about
references to a dynamic variable that doesn't yet exist.

Re-order restagraph.asd to remove a conflict, while we're at it. Define
the dispatchers _after_ the config, so that `*restagraph-acceptor*` is
already present, and the compiler stops complaining about it.