Modify get-relationship to return 'incoming-rels from Neo4j
In the method `get-relationship` as specialised on a Neo4j database,
return an 'incoming-rels instance instead of mistakenly trying to create
a 'schema-rels instance.

The problem with creating a `schema-rels instance is that it expects the
`target-type` argument to be in the form of a `schema-rtypes` instance,
which may not be feasible, and is not necessary under the circumstances.