Bugfix creation of non-existent resourcetypes
Check before anything else whether we even _have_ that lever. I mean,
that resourcetype. Raise a `client-error` if we don't, and leave it for
the handler to handle.

Internal tests: implement `restagraph-test::resources-basic`, to confirm
from the inside that we _can_ create instances of valid resourcetypes,
and _cannot_ create instances of invalid ones.

Client tests: augment `TestResources` with `test_fail_to_create_invalid_resourcetype`
as a basic regression test for this case.

Fixes #52