Bugfix: properly enforce relationship constraints
On creating a new relationship via POST, check the
relationship+destination pair for validity instead of only checking that
the source resourcetype has that kind of relationship to "something".

Add client tests.

Bugfix the RAW API dispatcher's call to `delete-relationship-by-path` -
as discovered while testing the above fix, the dispatcher was passing
the arguments in the wrong order.