Allow wildcard UIDs in URI paths
Augment `uri-node-helper` to skip the UID criterion for the last
resource on a path ending in a relationship.
This enables `RGinbound` criteria to filter for, e.g. all tags in use by
resources of a given type. It probably enables other sorts of mischief
as well.

Add a basic client test for this.