Return 200/Updated for PUT requests to the Raw API
This was previously returning 204 and an empty string, but that lacks
positive feedback that clients can use.
That also didn't match the semantics as defined in the Working Group
doc: https://httpwg.org/specs/rfc7231.html#PUT
...and it wasn't consistent with the schema API, either.

Update `README.md` and `docs/Demo_session.md` to make this explicit, and
to describe the reasoning behind it.

Also update the test suite for HTTP clients.