Add HTML output for GET /schema/v1
GET requests to `/schema/v1?format=html` will now result in an
HTML-formatted version, as found in the documentation, instead of the
default JSON version.

There is no equivalent for requests for a single resourcetype, because
at this point it really only makes sense if you're fetching the whole
thing.