Re-implement sanitise-uid to strip unreserved chars
Addresses #65 - stacktraces in response to apostrophes in the requested
URL in POST requests.

Switch from a deny-listing approach to an allow-listing one, based on
the list of unreserved characters in RFC 3986.
This is stricter than the previous policy, but should result in
better-quality URIs.

As of this commit, it has no tests.