Update process-filter's handling of inbound wildcards
When `RGinbound` is detected, use a custom version of `get-uri-parts`
that sanitises everything _except_ the asterisk wildcard character, _if_
it's found where a UID should be.

This is to catch a corner-case from the stricter version of
`sanitise-uid` (used by `get-uri-parts`) that came out of #65.

Update both internal and client-side testing to cover this. Client-side
tests that were involved:
- add TestSanitiseUid
- expand TestBasicResourceErrors

Clarify URL-safe in `README.md` and `docs/HTTP_API.md`