Continue using 'values' for enum attributes
Accept the keyword "values" in schema-definition files, for defining the
acceptable values for an enum-type varchar.
Also serialise `attrvalues` as "values", to be consistent about this.

This is a case where the internal and external implementations are
subject to sufficiently different pressures that it looks worth
maintaining this kind of inconsistency.